+++ dbos_transact/scheduler/scheduler.py
@@ -2,7 +2,7 @@ import threading
 from datetime import datetime, timezone
 from typing import Callable
 
-from .context import SetWorkflowUUID
+from ..context import SetWorkflowUUID
 from .croniter import croniter  # type: ignore
 
 ScheduledWorkflow = Callable[[datetime, datetime], None]
